CS558/HW1/requirements.txt
numpy
opencv-python
Pillow
argparse